feat: add registry testing scripts and update package.json
- Introduced `test:registry` and `test:install` scripts in package.json for automated testing of registry URLs and installation processes.
- Added GitHub Actions workflow for automated testing of registry URLs on push and pull request events.
- Implemented `test-registry-urls.mjs` for validating registry item URLs and `test-install-e2e.mjs` for end-to-end installation testing.
- Created necessary fixture files and components for testing purposes.
