+++ src/components/niko-table/components/data-table-faceted-filter.tsx
@@ -1,7 +1,7 @@
 "use client"
 
 import * as React from "react"
-import type { Table, Row } from "@tanstack/react-table"
+import type { Table } from "@tanstack/react-table"
 import {
   TableFacetedFilter,
   TableFacetedFilterContent,
@@ -13,71 +13,7 @@ import type { Option } from "../types"
 import { useDerivedColumnTitle } from "../hooks/use-derived-column-title"
 import { useGeneratedOptionsForColumn } from "../hooks/use-generated-options"
 import { formatLabel } from "../lib/format"
-
-/**
- * Get filtered rows excluding a specific column's filter.
- * This is useful when generating options for a column - we want to see
- * options that exist in the filtered dataset (from other filters) but
- * not be limited by the current column's own filter.
- */
-function getFilteredRowsExcludingColumn<TData>(
-  table: Table<TData>,
-  excludeColumnId: string,
-  columnFilters: Array<{ id: string; value: unknown }>,
-  globalFilter: unknown,
-): Row<TData>[] {
-  // Filter out the current column's filter
-  const otherFilters = columnFilters.filter(
-    filter => filter.id !== excludeColumnId,
-  )
-
-  // Get all core rows
-  const coreRows = table.getCoreRowModel().rows
-
-  // If no filters to apply (excluding the current column), return core rows
-  if (otherFilters.length === 0 && !globalFilter) {
-    return coreRows
-  }
-
-  // Filter rows manually, excluding the current column's filter
-  return coreRows.filter(row => {
-    // Apply column filters (excluding the current column)
-    for (const filter of otherFilters) {
-      const column = table.getColumn(filter.id)
-      if (!column) continue
-
-      const filterValue = filter.value
-      const filterFn = column.columnDef.filterFn || "extended"
-
-      // Skip if filter function is a string (built-in) and we don't have access
-      if (typeof filterFn === "string") {
-        // Use the table's filterFns
-        const fn = table.options.filterFns?.[filterFn]
-        if (fn && typeof fn === "function") {
-          if (!fn(row, filter.id, filterValue, () => {})) {
-            return false
-          }
-        }
-      } else if (typeof filterFn === "function") {
-        if (!filterFn(row, filter.id, filterValue, () => {})) {
-          return false
-        }
-      }
-    }
-
-    // Apply global filter if present
-    if (globalFilter) {
-      const globalFilterFn = table.options.globalFilterFn
-      if (globalFilterFn && typeof globalFilterFn === "function") {
-        if (!globalFilterFn(row, "global", globalFilter, () => {})) {
-          return false
-        }
-      }
-    }
-
-    return true
-  })
-}
+import { getFilteredRowsExcludingColumn } from "../lib/filter-rows"
 
 type DataTableFacetedFilterProps<TData, TValue> = Omit<
   TableFacetedFilterProps<TData, TValue>,
@@ -182,6 +118,14 @@ function useFacetedOptions<TData>({
   const columnFilters = state.columnFilters
   const globalFilter = state.globalFilter
 
+  /**
+   * REACTIVITY FIX: Extract coreRows outside memos so that when async data
+   * arrives, the new rows array reference triggers memo recomputation.
+   * Without this, `table` reference is stable across data changes and memos
+   * would return stale (empty) results after initial render with no data.
+   */
+  const coreRows = table.getCoreRowModel().rows
+
   // Fallback generator that works for any variant (text/boolean/etc.) to preserve
   // the original behavior of faceted filter for quick categorical filtering.
   const fallbackGenerated = React.useMemo((): Option[] => {
@@ -197,11 +141,12 @@ function useFacetedOptions<TData>({
     const rows = limitToFilteredRows
       ? getFilteredRowsExcludingColumn(
           table,
+          coreRows,
           accessorKey,
           columnFilters,
           globalFilter,
         )
-      : table.getCoreRowModel().rows
+      : coreRows
 
     const valueCounts = new Map<string, number>()
 
@@ -228,6 +173,7 @@ function useFacetedOptions<TData>({
     column,
     limitToFilteredRows,
     showCounts,
+    coreRows,
     table,
     columnFilters,
     globalFilter,
@@ -242,6 +188,7 @@ function useFacetedOptions<TData>({
         // We reuse fallbackGenerated's logic of getting occurrenceMap from rows
         const rows = getFilteredRowsExcludingColumn(
           table,
+          coreRows,
           accessorKey,
           columnFilters,
           globalFilter,
@@ -266,6 +213,7 @@ function useFacetedOptions<TData>({
     fallbackGenerated,
     limitToFilteredRows,
     column,
+    coreRows,
     table,
     accessorKey,
     columnFilters,
