+++ src/components/niko-table/hooks/use-generated-options.ts
@@ -1,76 +1,12 @@
 "use client"
 
 import * as React from "react"
-import type { Table, Row } from "@tanstack/react-table"
+import type { Table } from "@tanstack/react-table"
 
 import type { Option } from "../types"
 import { formatLabel } from "../lib/format"
 import { FILTER_VARIANTS } from "../lib/constants"
-
-/**
- * Get filtered rows excluding a specific column's filter.
- * This is useful when generating options for a column - we want to see
- * options that exist in the filtered dataset (from other filters) but
- * not be limited by the current column's own filter.
- */
-function getFilteredRowsExcludingColumn<TData>(
-  table: Table<TData>,
-  excludeColumnId: string,
-  columnFilters: Array<{ id: string; value: unknown }>,
-  globalFilter: unknown,
-): Row<TData>[] {
-  // Filter out the current column's filter
-  const otherFilters = columnFilters.filter(
-    filter => filter.id !== excludeColumnId,
-  )
-
-  // Get all core rows
-  const coreRows = table.getCoreRowModel().rows
-
-  // If no filters to apply (excluding the current column), return core rows
-  if (otherFilters.length === 0 && !globalFilter) {
-    return coreRows
-  }
-
-  // Filter rows manually, excluding the current column's filter
-  return coreRows.filter(row => {
-    // Apply column filters (excluding the current column)
-    for (const filter of otherFilters) {
-      const column = table.getColumn(filter.id)
-      if (!column) continue
-
-      const filterValue = filter.value
-      const filterFn = column.columnDef.filterFn || "extended"
-
-      // Skip if filter function is a string (built-in) and we don't have access
-      if (typeof filterFn === "string") {
-        // Use the table's filterFns
-        const fn = table.options.filterFns?.[filterFn]
-        if (fn && typeof fn === "function") {
-          if (!fn(row, filter.id, filterValue, () => {})) {
-            return false
-          }
-        }
-      } else if (typeof filterFn === "function") {
-        if (!filterFn(row, filter.id, filterValue, () => {})) {
-          return false
-        }
-      }
-    }
-
-    // Apply global filter if present
-    if (globalFilter) {
-      const globalFilterFn = table.options.globalFilterFn
-      if (globalFilterFn && typeof globalFilterFn === "function") {
-        if (!globalFilterFn(row, "global", globalFilter, () => {})) {
-          return false
-        }
-      }
-    }
-
-    return true
-  })
-}
+import { getFilteredRowsExcludingColumn } from "../lib/filter-rows"
 
 export interface GenerateOptionsConfig {
   /**
@@ -135,6 +71,14 @@ export function useGeneratedOptions<TData>(
    * IMPACT: Prevents unnecessary option regeneration when columns are stable.
    */
   const columns = React.useMemo(() => table.getAllColumns(), [table])
+
+  /**
+   * REACTIVITY FIX: Extract coreRows outside memos so that when async data
+   * arrives, the new rows array reference triggers memo recomputation.
+   * Without this, `table` reference is stable across data changes and memos
+   * would return stale (empty) results after initial render with no data.
+   */
+  const coreRows = table.getCoreRowModel().rows
 
   // Normalize array deps to stable strings for React hook linting
   const includeKey = includeColumns?.join(",") ?? ""
@@ -201,20 +145,22 @@ export function useGeneratedOptions<TData>(
       const optionSourceRows = limitToFilteredRows
         ? getFilteredRowsExcludingColumn(
             table,
+            coreRows,
             colId,
             columnFilters,
             globalFilter,
           )
-        : table.getCoreRowModel().rows
+        : coreRows
 
       const countSourceRows = colDynamicCounts
         ? getFilteredRowsExcludingColumn(
             table,
+            coreRows,
             colId,
             columnFilters,
             globalFilter,
           )
-        : table.getCoreRowModel().rows
+        : coreRows
 
       // If we have static options with augment strategy, we use static options and only calculate counts
       if (meta.options && meta.options.length > 0 && colMerge === "augment") {
@@ -322,6 +268,7 @@ export function useGeneratedOptions<TData>(
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [
     columns,
+    coreRows,
     table,
     dynamicCounts,
     showCounts,
